[core] Isolate contains refinement evaluation
